refactor(layout): split Navbar sub-components; fix setState-in-effect; restore hover-dismiss; move BrandIcon to icons/
