+++ src/components/icons/brand-icon.tsx
